Implemented a way to wait for transactions (and mark them as rollbackOnly) for shutdown

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/transaction/xaframework/XaLogicalLog.java
@@ -1457,6 +1457,7 @@ public class XaLogicalLog
      */
     public synchronized long rotate() throws IOException
     {
+        if ( writeBuffer.getFileChannelPosition() == LogIoUtils.LOG_HEADER_SIZE ) return xaTf.getLastCommittedTx();
         xaTf.flushAll();
         String newLogFile = getLog2FileName();
         String currentLogFile = getLog1FileName();
